[SDR-357] refactor: CommonHeader 폴더명 변경
