+++ fe/src/pages/ReviewList/ReviewList.tsx
@@ -1,5 +1,5 @@
 import Feeds from 'components/Common/Feeds/Feeds';
-import CommonHeader from 'components/Common/Header/CommonHeader';
+import CommonHeader from 'components/Common/CommonHeader/CommonHeader';
 import useAuth from 'hooks/useAuth';
 import useReviews from 'hooks/useReviews';
 import { useEffect } from 'react';
